backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
gui-agents>=0.1.2
pyautogui
e2b-desktop>=1.0.0
websockets>=12.0
